test(cms): loop_setting_propagation_spec.rb を既存 snippet spec に統合
propagation 用に独立ファイルを維持する必要性は薄いため、内容を責務の近い
既存 E2E 仕様に吸収してファイルを削除する。振り分けは以下の通り:

- node/liquid_snippet_spec.rb に移行:
  - SHIRASAGI / Liquid の記事フォルダ propagation (shared_examples)
  - 紐付け後に loop_setting を closed に変更しても公開ページが描画継続
  - 記事フォルダ編集画面のセレクタ UI で closed が除外される
- form/forms/liquid_snippet_spec.rb に移行:
  - Cms::Form layout html の loop_setting 経由 propagation
  - 定型フォームのレイアウトHTMLセレクタで closed/shirasagi が除外される

サイトドメインへの訪問が必要なテストは独自の around で Capybara.app_host を
確実に復元し、静的HTMLの干渉を防ぐため before で site.path を毎回クリーン
アップする (元ファイルの挙動を踏襲)。
